+++ firmware/common/Motor/MonsterShield.h
@@ -13,7 +13,7 @@
 class MonsterShield : public Motor
 {
   public:
-  MonsterShield(PWMDriver* _driver, uint32_t _channel,
+  MonsterShield(PWMDriver* _driver, uint8_t _channel,
                 bool _isComplementaryChannel, stm32_gpio_t* _CWGpio,
                 const uint32_t _CWPin, stm32_gpio_t* _CCWGpio, const uint32_t _CCWPin);
   virtual void changeDirection(eDirection _direction) override;
